+++ AbaTres.cs
@@ -1,0 +1,21 @@
+using System.Text.RegularExpressions;
+using System;
+using System.IO;
+using System.Linq;
+using System.Drawing;
+using System.Threading;
+using System.Diagnostics;
+using System.Windows.Forms;
+using System.ComponentModel;
+using System.Collections.Generic;
+
+namespace FormsComponentes
+{
+    public partial class AbaTres : TabPage
+    {
+        public AbaTres()
+        {
+            InitializeComponent();
+        }
+    }
+}
